(userlist_clnt_lookup_cookie) fixes for 64-bit platforms.

+++ userlist_clnt/lookup_cookie.c
@@ -32,7 +32,7 @@ userlist_clnt_lookup_cookie(struct userlist_clnt *clnt,
   struct userlist_pk_login_ok * answer = 0;
   void *void_answer = 0;
   int len;
-  int anslen;
+  size_t anslen;
   int res;
   int r;
